convertor-server/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.4
html5lib==1.0.1
idna==2.8
image==1.5.33
ipaddr==2.2.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.7.0
lockfile==0.12.2
msgpack==0.6.2
nltk==3.5
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
progress==1.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.3
pyparsing==2.4.6
pytesseract==0.3.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.4
pyzmq==20.0.0
regex==2020.11.13
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
urllib3==1.25.8
Wand==0.6.5
wcwidth==0.2.5
webencodings==0.5.1
